# docs/openapi.yaml
openapi: 3.1.0
info:
  title: FreshMart Agentic Analytics API
  description: "\n# FreshMart Agentic Analytics API\n\nWelcome to the **FreshMart\
    \ Agentic Analytics Platform** enterprise API documentation.\n\nFreshMart Agentic\
    \ Analytics provides secure, AI-powered natural-language analytics for FreshMart\
    \ enterprise business data across retail operations, customer relationship management,\
    \ human resources, inventory warehousing, purchasing, and financial performance.\n\
    \n---\n\n### Core Platform Capabilities\n\n* **AI-Powered Natural-Language Analytics:**\
    \ Inquire in plain English (e.g., *\"Why did profit decrease in August?\"* or\
    \ *\"Show monthly sales trend for the last 12 months\"*).\n* **LangGraph Multi-Agent\
    \ Orchestration:** Specialized agents collaborate autonomously:\n  * **Intent\
    \ Classifier:** Detects analytical goal, business domain, and inquiry category.\n\
    \  * **Permission Checker:** Enforces database table and column-level RBAC prior\
    \ to query generation.\n  * **Schema Planner:** Dynamically inspects schemas using\
    \ the Model Context Protocol (MCP).\n  * **SQL Generator & Corrector:** Synthesizes\
    \ read-only SQL via local `qwen2.5-coder:7b`.\n  * **SQL Validator:** Rigorous\
    \ AST parser preventing non-SELECT, destructive, or unauthorized queries.\n  *\
    \ **SQL Executor:** Safely runs validated SQL through MCP with row bounding and\
...
    \ `admin` / `AdminPassword123!`\n"
  contact:
...
paths:
  /api/v1/auth/login:
    post:
      tags:
      - Authentication
      summary: Authenticate User & Generate JWT Access Token
      description: 'Authenticates a FreshMart user with username and password, returning
        a signed JWT Bearer access token along with the user profile, role, and granted
        RBAC permission codes.


        **Supported Content-Types:**

        - `application/json`: Standard REST API clients

        - `application/x-www-form-urlencoded`: Interactive Swagger UI Authorize dialog
        and HTML forms


        **Available Demo Accounts (Password: `<Role>Password123!` or `AdminPassword123!`):**

        - `ceo`: Full enterprise analytics access

        - `sales.manager`: Sales, CRM, and inventory analytics

        - `finance.manager`: Financial P&L, expenses, and profit analytics

        - `hr.manager`: HRMS directory and employee salary analytics

        - `inventory.manager`: Products, warehouse stock, and purchasing

        - `erp.manager`: Integrated procurement, stock, sales, and HR

        - `admin`: System administration (user and role management)'
      operationId: loginUser
      requestBody:
        description: FreshMart application credentials submitted as JSON or Form URL-encoded
          data.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            examples:
              ceo_login:
                summary: CEO Account (Full Access)
                description: Login as CEO with full enterprise-wide visibility.
                value:
                  username: ceo
                  password: CeoPassword123!
              sales_manager_login:
                summary: Sales Manager Account
                description: Login as Sales Manager with sales, CRM, and inventory
                  permissions.
                value:
                  username: sales.manager
                  password: SalesPassword123!
              finance_manager_login:
                summary: Finance Manager Account
                description: Login as Finance Manager with P&L, profit, and expense
                  access.
                value:
                  username: finance.manager
                  password: FinancePassword123!
              hr_manager_login:
                summary: HR Manager Account
                description: Login as HR Manager with employee salary inspection rights.
                value:
                  username: hr.manager
                  password: HrPassword123!
              admin_login:
                summary: System Administrator
                description: Login as Technical Administrator for user and role management.
                value:
                  username: admin
                  password: AdminPassword123!
          application/x-www-form-urlencoded:
            schema:
              properties:
                username:
                  type: string
                  description: Application username (e.g., ceo, sales.manager, admin)
                  example: ceo
                password:
                  type: string
                  format: password
                  description: Application password
                  example: CeoPassword123!
                grant_type:
                  type: string
                  default: password
              type: object
              required:
              - username
              - password
        required: true
      responses:
        '200':
          description: User successfully authenticated. JWT access token and user
            profile returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          description: 'Bad Request: Missing username or password, or malformed request
            payload.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '401':
          description: 'Unauthorized: Incorrect username or password.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '403':
          description: 'Forbidden: User account is inactive or disabled by system
            administrator.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '422':
          description: 'Validation Error: Request payload could not be processed.'
  /api/v1/auth/me:
    get:
      tags:
      - Authentication
      summary: Get Current Authenticated User Profile
      description: Returns profile information, role metadata, and granted RBAC permission
        codes for the currently authenticated user identified by the verified JWT
        access token.
      operationId: getCurrentUserProfile
      responses:
        '200':
          description: User profile and granted permissions successfully retrieved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          description: 'Unauthorized: Missing, invalid, or expired JWT access token.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '403':
          description: 'Forbidden: User account is inactive or disabled.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
      security:
      - bearerAuth: []
      - OAuth2PasswordBearer: []
  /api/v1/admin/users:
    get:
      tags:
      - User Management
      summary: List All Registered Application Users
      description: 'Retrieves a complete list of all FreshMart application user accounts,
        including their assigned role, granted permission codes, workforce employee
        linkage, and active status.


        **Authorization:** Requires `MANAGE_USERS` permission (granted to `ADMIN`).'
      operationId: listAllApplicationUsers
      responses:
        '200':
          description: List of all user accounts successfully retrieved.
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/UserResponse'
                type: array
                title: Response 200 Listallapplicationusers
        '401':
          description: 'Unauthorized: Missing, invalid, or expired JWT access token.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '403':
          description: 'Forbidden: User lacks the required ''MANAGE_USERS'' permission.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
      security:
      - bearerAuth: []
      - OAuth2PasswordBearer: []
    post:
      tags:
      - User Management
      summary: Provision New Application User Account
      description: 'Provisions a new FreshMart user account with secure bcrypt password
        hashing and associates it with an existing application role.


        **Authorization:** Requires `MANAGE_USERS` permission (granted to `ADMIN`).'
      operationId: createApplicationUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
        required: true
      responses:
        '201':
          description: User account created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: 'Bad Request: Username or email already exists in system.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '401':
          description: 'Unauthorized: Missing, invalid, or expired JWT access token.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '403':
          description: 'Forbidden: User lacks the required ''MANAGE_USERS'' permission.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '404':
          description: 'Not Found: The specified role_id does not exist.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '422':
          description: 'Validation Error: Invalid payload attributes.'
      security:
      - bearerAuth: []
      - OAuth2PasswordBearer: []
...
  /api/v1/admin/roles:
    get:
      tags:
      - Role Management
      summary: List Application Roles & Granted Permission Codes
      description: 'Returns all 7 FreshMart application roles (ADMIN, CEO, HR_MANAGER,
        ERP_MANAGER, SALES_MANAGER, INVENTORY_MANAGER, FINANCE_MANAGER) along with
        their granular permission codes.


        **Authorization:** Requires `MANAGE_ROLES` permission (granted to `ADMIN`).'
      operationId: listApplicationRolesAndPermissions
      responses:
        '200':
          description: Roles and assigned permission codes successfully retrieved.
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/RoleResponse'
                type: array
                title: Response 200 Listapplicationrolesandpermissions
        '401':
          description: 'Unauthorized: Missing, invalid, or expired JWT access token.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '403':
          description: 'Forbidden: User lacks the required ''MANAGE_ROLES'' permission.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
      security:
      - bearerAuth: []
      - OAuth2PasswordBearer: []
...
  /api/v1/analytics/query:
    post:
      tags:
      - Agentic Analytics
      summary: Run Natural-Language Business Analytics Query
      description: 'Accepts a natural-language business analytics question and processes
        it through the authenticated LangGraph multi-agent pipeline with Model Context
        Protocol (MCP) database access.


        ### Processing Workflow:

        1. **Identity & RBAC Verification:** Authenticates user identity and enforces
        domain permissions (e.g., `VIEW_SALES`, `VIEW_INVENTORY`, `VIEW_FINANCE`,
        `VIEW_EMPLOYEE_SALARY`). Unauthorized inquiries are immediately rejected with
        HTTP 403 Forbidden before reaching AI or database layers.

        2. **Dynamic Schema Discovery:** Queries FreshMart database schema through
        MCP `describe_table` tool.

        3. **Read-Only SQL Generation:** Local Qwen 2.5-Coder model synthesizes precision
        analytical SQL.

        4. **Security Validation:** AST parser and regex validator reject non-SELECT,
        destructive, or unauthorized SQL.

        5. **MCP Tool Execution:** Executes SQL strictly through read-only MCP database
        interface with row bounds.

        6. **Multi-Agent Synthesis:** Llama 3.1 8B synthesizes grounded executive
        answers, calculates root-cause factors (for variance inquiries), extracts
        data provenance citations, and recommends chart visualization types.'
      operationId: runNaturalLanguageAnalyticsQuery
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnalyticsQueryRequest'
            examples:
              sales_trend:
                summary: Sales Trend (12 Months)
                description: Analyzes historical sales orders and returns monthly
                  revenue velocity.
                value:
                  question: Show monthly sales trend for the last 12 months.
              low_stock:
                summary: Inventory Status (Low Stock)
                description: Inspects warehouse inventory levels below product reorder
                  thresholds.
                value:
                  question: Which products are currently low in stock?
              cross_domain:
                summary: Cross-Domain (Revenue + Inventory)
                description: Combines sales order line items with inventory balances
                  across business databases.
                value:
                  question: Which products generated the highest revenue and are currently
                    low in stock?
              root_cause:
                summary: Root-Cause Analysis (August Profit Drop)
                description: Multi-agent causal diagnostic tracing net margin compression
                  in August 2026 to emergency freight and produce inflation.
                value:
                  question: Why did profit decrease in August?
              hrms_salary:
                summary: HRMS (Salaries by Department)
                description: Workforce analysis calculating department compensation
                  (Requires VIEW_EMPLOYEE_SALARY).
                value:
                  question: Show average employee salary by department.
        required: true
      responses:
        '200':
          description: Natural-language query successfully planned, executed, and
            synthesized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyticsQueryResponse'
        '400':
          description: 'Bad Request: Empty question, invalid parameters, or question
            too short.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '401':
          description: 'Unauthorized: Missing, invalid, or expired JWT access token.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '403':
          description: 'Forbidden: Access Denied due to insufficient RBAC domain permissions
            (e.g. attempting to view employee salary without VIEW_EMPLOYEE_SALARY).'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '422':
          description: 'Validation Error: Request payload failed schema validation.'
        '500':
          description: 'Internal Server Error: Pipeline execution or SQL synthesis
            failed.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
      security:
      - bearerAuth: []
      - OAuth2PasswordBearer: []
...
  /api/v1/analytics/insights:
    get:
      tags:
      - Executive Dashboard
      summary: Get Enterprise Insights, Anomalies & Strategic Recommendations
      description: 'Returns live structured FreshMart enterprise insights, anomalies,
        and recommendations strictly grounded in MySQL ledger data for the Insights
        dashboard.


        Includes automated detection of the August 2026 profit contraction, emergency
        air freight spikes, and warehouse safety stock deficit alerts.'
      operationId: getEnterpriseInsightsAndAnomalies
      responses:
        '200':
          description: Enterprise insights, statistical anomalies, and recommendations
            retrieved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnterpriseInsightsResponse'
        '401':
          description: 'Unauthorized: Missing, invalid, or expired JWT access token.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
      security:
      - bearerAuth: []
      - OAuth2PasswordBearer: []
...
components:
  schemas:
...
    AnalyticsQueryRequest:
      properties:
        question:
          type: string
          maxLength: 2000
          minLength: 1
          title: Question
          description: Natural-language business analytics question.
          example: Show monthly sales trend for the last 12 months.
        conversation_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Conversation Id
          description: Optional conversation session ID for contextual conversational
            memory across turns.
          example: session-marketing-01
      type: object
      required:
      - question
      title: AnalyticsQueryRequest
      description: Natural-language business question submitted to the multi-agent
        analytics pipeline.
      examples:
      - question: Show monthly sales trend for the last 12 months.
      - question: Which products are currently low in stock?
      - question: Which products generated the highest revenue and are currently low
          in stock?
      - question: Why did profit decrease in August?
      - question: Show average employee salary by department.
...
    EnterpriseInsightsResponse:
      properties:
        success:
          type: boolean
          title: Success
          description: Query execution status
          default: true
          example: true
        insights:
          items:
            $ref: '#/components/schemas/EnterpriseInsightCard'
          type: array
          title: Insights
          description: List of synthesized enterprise insights
        anomalies:
          items:
            $ref: '#/components/schemas/EnterpriseAnomalyCard'
          type: array
          title: Anomalies
          description: List of detected business anomalies
        timestamp:
          type: string
          format: date-time
          title: Timestamp
          description: UTC timestamp of telemetry calculation
      type: object
      required:
      - insights
      title: EnterpriseInsightsResponse
      description: Structured enterprise insights, anomalies, and recommendations
        grounded in MySQL ledger data.
    HTTPError:
      properties:
        detail:
          type: string
          title: Detail
          description: Human-readable explanation of why the request failed.
          example: 'Access Denied: You do not have permission ''VIEW_EMPLOYEE_SALARY''.'
      type: object
      required:
      - detail
      title: HTTPError
      description: Standard RFC 7807-compatible error response detail.
...
    RoleResponse:
      properties:
        role_id:
          type: integer
          title: Role Id
          description: Unique role identifier
          example: 5
        role_name:
          type: string
          title: Role Name
          description: Unique role name
          example: SALES_MANAGER
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
          description: Role purpose description
          example: Sales Manager. Customers, CRM, orders, and sales trends.
        permissions:
          items:
            type: string
          type: array
          title: Permissions
          description: List of permission codes assigned to this role
          example:
          - VIEW_CRM
          - VIEW_SALES
          - VIEW_INVENTORY
      type: object
      required:
      - role_id
      - role_name
      - permissions
      title: RoleResponse
      description: Application role representation with assigned permissions.
...
    TokenResponse:
      properties:
        access_token:
          type: string
          title: Access Token
          description: Signed JWT Bearer access token containing claims and permissions
        token_type:
          type: string
          title: Token Type
          description: Token protocol type
          default: bearer
          example: bearer
        user:
          $ref: '#/components/schemas/UserProfile'
          description: Authenticated user profile and permissions
      type: object
      required:
      - access_token
      - user
      title: TokenResponse
      description: JWT token response returned upon successful authentication.
...
    UserCreate:
      properties:
        username:
          type: string
          maxLength: 60
          minLength: 3
          title: Username
          description: Unique username for the new account
          example: anita.finance
        email:
          type: string
          title: Email
          description: Unique email address
          example: anita.f@freshmart.local
        password:
          type: string
          minLength: 6
          title: Password
          description: Initial plain text password (hashed securely with bcrypt)
          example: SecurePass2026!
        full_name:
          type: string
          maxLength: 120
          title: Full Name
          description: Full display name of the user
          example: Anita Sharma
        role_id:
          type: integer
          title: Role Id
          description: 'Role ID to assign to this user (1: ADMIN, 2: CEO, 3: HR_MANAGER,
            4: ERP_MANAGER, 5: SALES_MANAGER, 6: INVENTORY_MANAGER, 7: FINANCE_MANAGER)'
          example: 7
        employee_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Employee Id
          description: Optional MySQL employee ID to associate
          example: E010
      type: object
      required:
      - username
      - email
      - password
      - full_name
      - role_id
      title: UserCreate
      description: Payload to provision a new application user account.
    UserProfile:
      properties:
        user_id:
          type: integer
          title: User Id
          description: Unique integer user identifier
          example: 2
        username:
          type: string
          title: Username
          description: Unique application username
          example: ceo
        email:
          type: string
          title: Email
          description: Corporate email address
          example: ceo@freshmart.local
        full_name:
          type: string
          title: Full Name
          description: User's full name
          example: Vikram Malhotra
        role:
          type: string
          title: Role
          description: Assigned application role name
          example: CEO
        employee_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Employee Id
          description: Linked MySQL workforce employee ID
          example: E006
        permissions:
          items:
            type: string
          type: array
          title: Permissions
          description: List of granular RBAC permission codes granted to this user
          example:
          - VIEW_HR
          - VIEW_EMPLOYEE_SALARY
          - VIEW_CRM
          - VIEW_SALES
          - VIEW_INVENTORY
          - VIEW_PURCHASES
          - VIEW_FINANCE
          - VIEW_PROFIT
          - VIEW_EXPENSES
        is_active:
          type: boolean
          title: Is Active
          description: Account active status flag
          default: true
          example: true
      type: object
      required:
      - user_id
      - username
      - email
      - full_name
      - role
      - permissions
      title: UserProfile
      description: Current authenticated user profile and granted permissions.
    UserResponse:
      properties:
        user_id:
          type: integer
          title: User Id
          description: Unique user identifier
          example: 1
        username:
          type: string
          title: Username
          description: Unique username
          example: admin
        email:
          type: string
          title: Email
          description: Email address
          example: admin@freshmart.local
        full_name:
          type: string
          title: Full Name
          description: Display name
          example: Technical Administrator
        role_id:
          type: integer
          title: Role Id
          description: Foreign key reference to assigned role
          example: 1
        role_name:
          type: string
          title: Role Name
          description: Name of assigned role
          example: ADMIN
        employee_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Employee Id
          description: Linked MySQL workforce ID
        is_active:
          type: boolean
          title: Is Active
          description: Whether user account is active
          example: true
        permissions:
          items:
            type: string
          type: array
          title: Permissions
          description: List of granular permission codes granted via role
          example:
          - MANAGE_USERS
          - MANAGE_ROLES
        created_at:
          type: string
          format: date-time
          title: Created At
          description: Account creation UTC timestamp
      type: object
      required:
      - user_id
      - username
      - email
      - full_name
      - role_id
      - role_name
      - is_active
      - permissions
      - created_at
      title: UserResponse
      description: Detailed user account representation.
...
    LoginRequest:
      description: User login request payload.
      examples:
      - password: CeoPassword123!
        username: ceo
      - password: SalesPassword123!
        username: sales.manager
      - password: HrPassword123!
        username: hr.manager
      - password: AdminPassword123!
        username: admin
      properties:
        username:
          description: Application username (e.g. ceo, sales.manager, admin)
          example: ceo
          title: Username
          type: string
        password:
          description: Application account password
          example: CeoPassword123!
          title: Password
          type: string
      required:
      - username
      - password
      title: LoginRequest
      type: object
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Authorization header using the Bearer scheme. Enter token directly
        or as 'Bearer <token>'.
    OAuth2PasswordBearer:
      type: oauth2
      description: 'OAuth2 Password Flow: Authenticate using username and password'
      flows:
        password:
          scopes: {}
          tokenUrl: /api/v1/auth/login
...